Rename SmsGatewayPluginIncomingEventInterface to SmsIncomingEventInterface
